requirements.txt
Flask==1.1.1
beautifulsoup4==4.8.2
line-bot-sdk==1.16.0
lxml==4.5.0